+++ next/src/components/questionnaires/form/QuestionnaireForm.tsx
@@ -3,21 +3,18 @@ import { useNavigate } from '@tanstack/react-router'
 import { Controller, type SubmitHandler, useForm } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 
-import Checkbox from '@/components/ui/form/Checkbox'
 import Field from '@/components/ui/form/Field'
 import Form from '@/components/ui/form/Form'
 import Input from '@/components/ui/form/Input'
-import Label from '@/components/ui/form/Label'
 import RadioGroup from '@/components/ui/form/RadioGroup'
 import {
   FlowLogics,
   FormulasLanguages,
   type Questionnaire,
-  TargetModes,
 } from '@/models/questionnaires'
 
+import SelectTargetMode from '../../ui/form/SelectTargetMode'
 import { type FormValues, schema } from './schema'
-import { changeSetValue } from './utils/set'
 
 type Props = {
   /** In an update case, initial questionnaire value. */
@@ -101,54 +98,12 @@ export default function QuestionnaireForm({
         control={control}
         rules={{ required: true }}
         render={({ field, fieldState: { error } }) => (
-          <>
-            <Label required>{t('questionnaire.common.targetMode')}</Label>
-            <div className="flex gap-x-4">
-              <Checkbox
-                label={'CAPI'}
-                checked={field.value?.has(TargetModes.CAPI)}
-                {...field}
-                onChange={(v) =>
-                  field.onChange(
-                    changeSetValue(field.value, TargetModes.CAPI, v),
-                  )
-                }
-              />
-              <Checkbox
-                label={'CAWI'}
-                checked={field.value?.has(TargetModes.CAWI)}
-                {...field}
-                onChange={(v) =>
-                  field.onChange(
-                    changeSetValue(field.value, TargetModes.CAWI, v),
-                  )
-                }
-              />
-              <Checkbox
-                label={'CATI'}
-                checked={field.value?.has(TargetModes.CATI)}
-                {...field}
-                onChange={(v) =>
-                  field.onChange(
-                    changeSetValue(field.value, TargetModes.CATI, v),
-                  )
-                }
-              />
-              <Checkbox
-                label={'PAPI'}
-                checked={field.value?.has(TargetModes.PAPI)}
-                {...field}
-                onChange={(v) =>
-                  field.onChange(
-                    changeSetValue(field.value, TargetModes.PAPI, v),
-                  )
-                }
-              />
-            </div>
-            {error ? (
-              <div className="text-sm text-error ml-1">{error.message}</div>
-            ) : null}
-          </>
+          <SelectTargetMode
+            value={field.value}
+            onChange={field.onChange}
+            multiple
+            error={error?.message}
+          />
         )}
       />
       <div>