+++ next/src/lib/i18n/locales/fr/translation.ts
@@ -3,6 +3,7 @@ import codesList from './codesList.json'
 import codesLists from './codesLists.json'
 import common from './common.json'
 import crumb from './crumb.json'
+import details from './details.json'
 import error from './error.json'
 import filter from './filter.json'
 import history from './history.json'
@@ -22,6 +23,7 @@ export const translation = {
   codesLists,
   common,
   crumb,
+  details,
   error,
   filter,
   history,